refactor: 承認系 entity を domain/entities/approval/ へ区画化し barrel/eslint パスを更新 (#421)
EstimateApplication / EstimateApprovalStep / EstimateApprovalExemption と
隣接テストを domain/entities/approval/ 配下へ集約。

- 移動本体の values 参照は +1 階層（../values → ../../values）、兄弟エンティティ
  ./EstimateApprovalStep と自テスト ../Subject は共移動のため不変。
- entities barrel の承認系 re-export 2行を ./approval/ へ更新。
- eslint no-restricted-imports の EstimateApprovalStep を新パスへ（1行）。
  子エンティティ参照許可オーバーライド(entities/** グロブ)は approval/ を自動カバー。

build/eslint(estimate)/test green。
