+++ src/server/subdomains/estimate/domain/entities/approval/__tests__/EstimateApprovalExemption.test.ts
@@ -1,9 +1,9 @@
 import { EmployeeId } from "@subdomains/employee/domain/values/EmployeeId";
 import { describe, expect, it } from "vitest";
 import { EstimateApprovalExemption } from "../EstimateApprovalExemption";
-import { EstimateApprovalExemptionId } from "../../values/approval/EstimateApprovalExemptionId";
-import { EstimateExemptionReason } from "../../values/approval/EstimateExemptionReason";
-import { EstimateVariationId } from "../../values/EstimateVariationId";
+import { EstimateApprovalExemptionId } from "../../../values/approval/EstimateApprovalExemptionId";
+import { EstimateExemptionReason } from "../../../values/approval/EstimateExemptionReason";
+import { EstimateVariationId } from "../../../values/EstimateVariationId";
 
 describe("EstimateApprovalExemption", () => {
   describe("create()", () => {
