+++ src/server/subdomains/estimate/domain/entities/index.ts
@@ -19,11 +19,11 @@
 export { Estimate } from "./Estimate";
 // EstimateApprovalExemption は承認免除の独立集約ルート（ADR-0054）。子エンティティを
 // 持たない薄い集約のため、ルートをそのまま公開する。
-export { EstimateApprovalExemption } from "./EstimateApprovalExemption";
+export { EstimateApprovalExemption } from "./approval/EstimateApprovalExemption";
 // EstimateApplication は見積申請の集約ルート。create ファクトリ（ADR-0036）は VO 計画
 // ApprovalChainPlan と identity VO のみを入力に取り、子エンティティ EstimateApprovalStep
 // 型は露出しないため、ルートとファクトリを公開しても境界規約（ADR-0027）を損なわない。
-export { EstimateApplication } from "./EstimateApplication";
+export { EstimateApplication } from "./approval/EstimateApplication";
 // EstimateFactory は集約外（アプリ層）からの集約生成口。入出力は VO 記述子と
 // 集約ルート Estimate のみで、子エンティティ型は露出しないため、本バレルから公開しても
 // 集約境界規約を損なわない。
